refactor: merge tests
